services/pyapi/requirements.txt
Flask
flask-cli
psycopg2-binary
flask-restful
pandas
paramiko
gunicorn
